@@ -3,6 +3,7 @@
   "version": "0.0.0",
   "description": "OpenAPI Explorer - API viewer with dynamically generated components, documentation, and interaction console",
   "author": "Rhosys Developers <developers@rhosys.ch>",
+  "type": "module",
   "repository": {
     "type": "git",
     "url": "https://github.com/Rhosys/openapi-explorer.git"
@@ -59,7 +60,7 @@
     "start": "webpack serve --mode=development --port=8080",
     "serve": "serve -p 8080 mocks",
     "lint": "eslint --ext .js src",
-    "test": "babel src -d lib && mocha -r esm tests/*.test.js"
+    "test": "ava tests/**/*.test.js"
   },
   "devDependencies": {
     "@babel/cli": "^7.13.14",
@@ -69,6 +70,7 @@
     "@babel/plugin-proposal-optional-chaining": "^7.18.9",
     "@babel/plugin-transform-modules-commonjs": "^7.18.6",
     "@babel/preset-env": "^7.13.12",
+    "ava": "^4.3.3",
     "babel-loader": "^8.2.2",
     "babel-plugin-template-html-minifier": "^4.1.0",
     "chai": "^4.2.0",
@@ -87,13 +89,13 @@
     "eslint-plugin-promise": "^5.1.0",
     "eslint-plugin-vue": "^7.17.0",
     "eslint-webpack-plugin": "2.5.3",
-    "esm": "^3.2.25",
     "file-loader": "^6.2.0",
     "filemanager-webpack-plugin": "^4.0.0",
     "fs-extra": "^8.1.0",
     "glob": "^7.1.6",
     "html-webpack-plugin": "^5.3.1",
     "inspectpack": "^4.7.1",
+    "json-loader": "^0.5.7",
     "mocha": "^7.2.0",
     "npm-run-all": "^4.1.5",
     "sinon": "^7.5.0",
@@ -106,6 +108,6 @@
     "webpack-dev-server": "^3.11.2"
   },
   "engines": {
-    "node": ">=12"
+    "node": ">=14"
   }
 }
