@@ -1,57 +1,55 @@
-const { expect } = require('chai');
-const { describe, it } = require('mocha');
-const xmlFormatter = require('../src/utils/xml/xml').default;
+import test from 'ava';
+import { expect } from 'chai';
 
-const schemaUtils = require('../src/utils/schema-utils');
+import xmlFormatter from '../src/utils/xml/xml.js';
+import { getExampleValuesFromSchema } from '../src/utils/schema-utils.js';
 
-describe('schema-utils.js', () => {
-  describe('getExampleValuesFromSchema', () => {
-    it('xml', () => {
-      const schema = {
+test('schema-utils.js getExampleValuesFromSchema', t => {
+  const schema = {
+    type: 'object',
+    required: ['name', 'photoUrls'],
+    properties: {
+      id: { type: 'integer', format: 'int64' },
+      category: {
         type: 'object',
-        required: ['name', 'photoUrls'],
         properties: {
           id: { type: 'integer', format: 'int64' },
-          category: {
-            type: 'object',
-            properties: {
-              id: { type: 'integer', format: 'int64' },
-              name: { type: 'string' },
-            },
-            xml: { name: 'Category' },
-            $$ref: 'https://petstore.swagger.io/v2/swagger.json#/definitions/Category',
-          },
-          name: { type: 'string', example: 'doggie' },
-          photoUrls: {
-            type: 'array',
-            xml: { wrapped: true },
-            items: { type: 'string', xml: { name: 'photoUrl' } },
-          },
-          tags: {
-            type: 'array',
-            xml: { wrapped: true },
-            items: {
-              type: 'object',
-              properties: {
-                id: { type: 'integer', format: 'int64' },
-                name: { type: 'string' },
-              },
-              xml: { name: 'Tag' },
-              $$ref: 'https://petstore.swagger.io/v2/swagger.json#/definitions/Tag',
-            },
-          },
-          status: {
-            type: 'string',
-            description: 'pet status in the store',
-            enum: ['available', 'pending', 'sold'],
-          },
+          name: { type: 'string' },
         },
-        xml: { name: 'Pet' },
-        $$ref: 'https://petstore.swagger.io/v2/swagger.json#/definitions/Pet',
-      };
-      const result = schemaUtils.getExampleValuesFromSchema(schema, { xml: true });
-      const formattedResult = xmlFormatter(result, { indent: '  ' });
-      const expectedFormattedResult = `<Pet>
+        xml: { name: 'Category' },
+        $$ref: 'https://petstore.swagger.io/v2/swagger.json#/definitions/Category',
+      },
+      name: { type: 'string', example: 'doggie' },
+      photoUrls: {
+        type: 'array',
+        xml: { wrapped: true },
+        items: { type: 'string', xml: { name: 'photoUrl' } },
+      },
+      tags: {
+        type: 'array',
+        xml: { wrapped: true },
+        items: {
+          type: 'object',
+          properties: {
+            id: { type: 'integer', format: 'int64' },
+            name: { type: 'string' },
+          },
+          xml: { name: 'Tag' },
+          $$ref: 'https://petstore.swagger.io/v2/swagger.json#/definitions/Tag',
+        },
+      },
+      status: {
+        type: 'string',
+        description: 'pet status in the store',
+        enum: ['available', 'pending', 'sold'],
+      },
+    },
+    xml: { name: 'Pet' },
+    $$ref: 'https://petstore.swagger.io/v2/swagger.json#/definitions/Pet',
+  };
+  const result = getExampleValuesFromSchema(schema, { xml: true });
+  const formattedResult = xmlFormatter(result, { indent: '  ' });
+  const expectedFormattedResult = `<Pet>
   <id>0</id>
   <Category>
     <id>0</id>
@@ -69,7 +67,6 @@ describe('schema-utils.js', () => {
   </tags>
   <status>available</status>
 </Pet>`;
-      expect(formattedResult).to.eql(expectedFormattedResult);
-    });
-  });
+  expect(formattedResult).to.eql(expectedFormattedResult);
+  t.pass();
 });
